+++ src/element/Grid.cpp
@@ -4,8 +4,8 @@
 using namespace sfSnake;
 
 Grid::Grid()
-    : gridVline(sf::Vector2f(Game::GlobalVideoMode.width / 500.0f, Game::GlobalVideoMode.height)),
-      gridHline(sf::Vector2f(Game::GlobalVideoMode.width, Game::GlobalVideoMode.width / 500.0f))
+    : gridVline(sf::Vector2f(Game::GlobalVideoMode.size.x / 500.0f, Game::GlobalVideoMode.size.y)),
+      gridHline(sf::Vector2f(Game::GlobalVideoMode.size.x, Game::GlobalVideoMode.size.x / 500.0f))
 {
 }
 
@@ -14,16 +14,16 @@ void Grid::render(sf::RenderWindow &window)
     gridVline.setFillColor(Game::Color::Grid[Game::GridColor]);
     gridHline.setFillColor(Game::Color::Grid[Game::GridColor]);
 
-    static float gridSpace = Game::GlobalVideoMode.width / 50.0f;
-    for (float i = 0.f; i < Game::GlobalVideoMode.width; i += gridSpace)
+    static float gridSpace = Game::GlobalVideoMode.size.x / 50.0f;
+    for (float i = 0.f; i < Game::GlobalVideoMode.size.x; i += gridSpace)
     {
-        gridVline.setPosition(i, 0);
+        gridVline.setPosition({i, 0.f});
         window.draw(gridVline);
     }
 
-    for (float i = 0.f; i < Game::GlobalVideoMode.height; i += gridSpace)
+    for (float i = 0.f; i < Game::GlobalVideoMode.size.y; i += gridSpace)
     {
-        gridHline.setPosition(0, i);
+        gridHline.setPosition({0.f, i});
         window.draw(gridHline);
     }
 }