+++ src/screen/GameOverScreen.cpp
@@ -12,36 +12,38 @@
 using namespace sfSnake;
 
 GameOverScreen::GameOverScreen(std::size_t score)
-    : button_(3), score_(score)
+    : text_(Game::GlobalFont), button_(3), score_(score)
 {
     text_.setFont(Game::GlobalFont);
     text_.setString(sf::String(L"最终分数：") + std::to_string(score));
     text_.setFillColor(Game::Color::Yellow);
-    text_.setCharacterSize(Game::GlobalVideoMode.width / 15);
+    text_.setCharacterSize(Game::GlobalVideoMode.size.x / 15);
     setOriginMiddle(text_);
-    text_.setPosition(Game::GlobalVideoMode.width / 2.0f, Game::GlobalVideoMode.height / 4.0f);
+    text_.setPosition(
+        {Game::GlobalVideoMode.size.x / 2.0f,
+         Game::GlobalVideoMode.size.y / 4.0f});
 
     button_[0].update("assets/image/optionUI.png");
     button_[1].update("assets/image/restartUI.png");
     button_[2].update("assets/image/exitUI.png");
 
-    button_[0].setPosition(Game::GlobalVideoMode.width / 3.0, Game::GlobalVideoMode.height / 5.0 * 3.0);
-    button_[1].setPosition(Game::GlobalVideoMode.width / 2.0, Game::GlobalVideoMode.height / 5.0 * 3.0);
-    button_[2].setPosition(Game::GlobalVideoMode.width / 3.0 * 2.0, Game::GlobalVideoMode.height / 5.0 * 3.0);
+    button_[0].setPosition(Game::GlobalVideoMode.size.x / 3.0f, Game::GlobalVideoMode.size.y / 5.0f * 3.0f);
+    button_[1].setPosition(Game::GlobalVideoMode.size.x / 2.0f, Game::GlobalVideoMode.size.y / 5.0f * 3.0f);
+    button_[2].setPosition(Game::GlobalVideoMode.size.x / 3.0f * 2.0f, Game::GlobalVideoMode.size.y / 5.0f * 3.0f);
 
     helpButton_.settings(
         L"帮助",
         Game::GlobalFont,
-        Game::GlobalVideoMode.width / 20.0,
+        Game::GlobalVideoMode.size.x / 20.0f,
         Game::Color::Green,
-        sf::Vector2f(Game::GlobalVideoMode.width / 5.0f * 2.0f, Game::GlobalVideoMode.height / 5.0f * 4.0f));
+        sf::Vector2f(Game::GlobalVideoMode.size.x / 5.0f * 2.0f, Game::GlobalVideoMode.size.y / 5.0f * 4.0f));
 
     aboutButton_.settings(
         L"关于",
         Game::GlobalFont,
-        Game::GlobalVideoMode.width / 20.0,
+        Game::GlobalVideoMode.size.x / 20.0f,
         Game::Color::Green,
-        sf::Vector2f(Game::GlobalVideoMode.width / 5.0f * 3.0f, Game::GlobalVideoMode.height / 5.0f * 4.0f));
+        sf::Vector2f(Game::GlobalVideoMode.size.x / 5.0f * 3.0f, Game::GlobalVideoMode.size.y / 5.0f * 4.0f));
 }
 
 void GameOverScreen::handleInput(sf::RenderWindow &window)
@@ -57,7 +59,7 @@ void GameOverScreen::handleInput(sf::RenderWindow &window)
     if (button_[0].contain(mousePosition))
     {
         button_[0].focused(true);
-        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Left))
+        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Button::Left))
         {
             Game::mouseButtonCDtime = sf::Time::Zero;
             Game::mouseButtonLocked = true;
@@ -70,7 +72,7 @@ void GameOverScreen::handleInput(sf::RenderWindow &window)
     if (button_[1].contain(mousePosition))
     {
         button_[1].focused(true);
-        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Left))
+        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Button::Left))
         {
             Game::mouseButtonCDtime = Game::mouseButtonClock.restart();
             Game::mouseButtonLocked = true;
@@ -82,7 +84,7 @@ void GameOverScreen::handleInput(sf::RenderWindow &window)
     if (button_[2].contain(mousePosition))
     {
         button_[2].focused(true);
-        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Left))
+        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Button::Left))
         {
             window.close();
             return;
@@ -92,7 +94,7 @@ void GameOverScreen::handleInput(sf::RenderWindow &window)
     if (helpButton_.contains(mousePosition))
     {
         helpButton_.focused();
-        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Left))
+        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Button::Left))
         {
             helpButton_.seleted();
             Game::mouseButtonCDtime = sf::Time::Zero;
@@ -106,7 +108,7 @@ void GameOverScreen::handleInput(sf::RenderWindow &window)
     if (aboutButton_.contains(mousePosition))
     {
         aboutButton_.focused();
-        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Left))
+        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Button::Left))
         {
             aboutButton_.seleted();
             Game::mouseButtonCDtime = sf::Time::Zero;