+++ src/screen/HelpScreen.cpp
@@ -11,6 +11,7 @@
 using namespace sfSnake;
 
 HelpScreen::HelpScreen()
+    : text_(Game::GlobalFont)
 { 
     if(!Game::ifShowedHelp) Game::ifShowedHelp = true;
 
@@ -22,24 +23,28 @@ HelpScreen::HelpScreen()
         sf::String(L"屏幕上有5种不同颜色的水果，它们分别对应于得分：\n") +
         sf::String(L"黑色、棕色（不得分），红色（3分），蓝色（2分），绿色（1分）\n") +
         sf::String(L"点击设置按钮，设置网格开关、网格颜色、背景颜色"));
-    text_.setCharacterSize(Game::GlobalVideoMode.width / 35.0f);
+    text_.setCharacterSize(Game::GlobalVideoMode.size.x / 35.0f);
     text_.setFillColor(Game::Color::Green);
     setOriginMiddle(text_);
-    text_.setPosition(Game::GlobalVideoMode.width / 2.0, Game::GlobalVideoMode.height / 5.0 * 3.0);
+    text_.setPosition(
+        {Game::GlobalVideoMode.size.x / 2.0f,
+         Game::GlobalVideoMode.size.y / 5.0f * 3.0f});
 
     returnButton_.update("assets/image/returnUI.png", 1 / 16.0f);
-    returnButton_.setPosition(Game::GlobalVideoMode.width / 15.0f, Game::GlobalVideoMode.width / 15.0f);
+    returnButton_.setPosition(
+        Game::GlobalVideoMode.size.x / 15.0f,
+        Game::GlobalVideoMode.size.x / 15.0f);
 }
 
 void HelpScreen::handleInput(sf::RenderWindow &window)
 {
-    auto mousePosition = sf::Mouse::getPosition(window);
+    const auto mousePosition = sf::Mouse::getPosition(window);
 
     returnButton_.focused(false);
     if (returnButton_.contain(mousePosition))
     {
         returnButton_.focused(true);
-        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Left))
+        if (!Game::mouseButtonLocked && sf::Mouse::isButtonPressed(sf::Mouse::Button::Left))
         {
             Game::mouseButtonCDtime = sf::Time::Zero;
             Game::mouseButtonLocked = true;